Add support for indexing into variables
This adds support for indexing into variables of the new type "List".
The syntax is typical C-style array syntax, so for example, the
following is valid for interpolating the element of the list in variable
foo at position 1:

    "${foo[1]}"

There is a restriction that lists must be homogenous in type. This is to
enable the type checker to work without having to actually perform the
indexing operation. During discussion there was no compelling use case
for lists of heterogenous types - this is open for discussion in future
though if one is found.