Improve error when concat output is not a string
This commit changes the error message returned when the arguments to
concat type check as something other than a string, to indicate that
it is the entire HIL expression that needs to result in a string. The
previous behaviour made debugging in consumers difficult when the
expression consisted of a function which took a string as an argument
and returned something other than a string.