Add support for indexing map variables
This adds support for indexing into variables of the new type "Map".
The syntax is typical Go-style map syntax, so for example, the
following is valid for interpolating the value of the map foo at key
"bar":

    `${foo["bar"]}`

There is a restriction that map values must be homogenous in type. This
is to enable the type checker to work without having to actually perform
the indexing operation. This is a similar restriction to that placedo n
lists.

During discussion there was no compelling use case for maps or lists of
heterogenous types - this is open for discussion in future though if one
is found.

This also expands out the range of test cases covered for lists during
type checking and evaluation to include empty lists.