Introduce Output nodes and EvaluationResult type
hil.Eval now returns a structure including the Value and Result of an
interpolation, in terms not associated with the ast. This prevents the
proliferation of ast.TypeX types in consumer code, and provides a
natural boundary between inside and outside hil.

We also change the output types for lists and maps to use the "natural"
Go types for these ([]interface{} and map[string]interface{}
respectively) which again prevents the proliferation of ast.Variable
processing code in consumers of the library.

This changes the signature - the old tests remain for internal code, and
there are some new ones added for the new interface.