Add bool type
This PR adds a new `TypeBool`. That allows us to write functions that
returns true or false. I think this is an important step for
https://github.com/hashicorp/terraform/issues/1604 as well

Here is an example `ast.Function` called `equal` that compares two
values and returns true or false (compiles with this PR).

https://play.golang.org/p/JpJ7GzA454

I've tried to check whether this breaks something, but right now it
passes all tests. Unfortunately I don't like the yacc parser, so it's
hard to understand it. Let me know if this is not suitable or if
something is wrong here.