ast.Index should visit target and key
The visitor pattern shouldn't discriminate what values are and are not
visited. This allows us to use the type checking stack and the eval
stack more properly and avoid evaling multiple nodes within the index
eval.

In turn, this enables all eval operations to work in both targets and
keys (fixes #28). We still enforce that the list target be a variable
access, however.