Support implicit conversion for list/map index key
HIL is a bit PHP-esque in that it does implicit string to number and
vice versa conversion where possible. For example `"2" - 1` equates to
1. This was a conscious decision made since without it, there ended up
being a LOT of explicit type conversion necessary since things like
counts in Terraform were/are strings.

This PR extends implicits to list and map index operations, allowing
variables that are potentially strings to be used as indexes without
type conversion operations (which don't exist right now, so you'd have
to do weird math tricks).