TypeUnknown for "unknown" values
This is a special type (and sentinel value for InterfaceToVariable)
where any operation against it results in a TypeUnknown result. You can
see the tests to see how that works.

This is nice because if you have something like a function, map, or list
that DOESN'T result in an unknown value then the evaluation will work,
but if you happen to access an unknown value at runtime, hil as a whole
will return the unknown value immediately.

This helps Terraform tremendously because it can stop special casing
unknown values and actual mark them as such as just give them to HIL for
normal evaluation all the time. If HIL returns TypeUnknown, then
Terraform knows the value is computed. If in Terraform this happens at
apply-time, then its an error (no computed values during apply).

This has zero effect on previous HIL programs: it is impossible to
arbitrarily introduce "TypeUnknown" unless explicitly passed in by the
HIL caller (not the HIL program).

And, other uses of HIL actually could very well have computed values
(thinking Nomad here) and now that HIL supports it 1st class we can just
magically use it.