Re-implement the HIL scanner as plain Go code
Previously HIL used a generated scanner and parser, but the generated
parser is finicky and hard to extend. A later commit will replace that
parser with a hand-written recursive descent parser, but as a first step
the scanner is hand-written so it can be used as input to this forthcoming
parser.

To start this scanner produces comparable output to the old scanner but
for the following exceptions:

- Multi-byte tokens are just returned verbatim as they were presented in
  the source input. It's the parser's responsibility to resolve any
  escape sequences in LITERAL and STRING tokens and to ensure that
  INT and FLOAT tokens are of the appropriate range and syntax.

- The "column" position is maintained in runes rather than bytes, so that
  UTF-8 sequences can be counted more properly. (This is still not 100%
  accurate, since combining forms will count as multiple runes, but
  closer than before.)

- Identifiers may contain any characters that Unicode considers to be
  letters or combining marks, so they can now e.g. include latin accent
  characters or letters in non-Latin scripts.

- Sequences like foo.bar.*.baz are handled as a sequence of separate
  IDENTIFIER, PERIOD and STAR tokens, which the parser should then stick
  back together to make a variable name to include in the AST. This means
  that the parser can detect incorrect sequences like "foo..bar" and
  raise a nicer error message for them than the scanner would be able to.

The tests from the old scanner are preserved and adapted to the above
changes. Some of the test cases are adjusted to exercise the UTF-8
handling and verify the correct handling of some invalid cases.