Replace yacc parser with hand-written Go parser
The yacc parser was non-reentrant, generated poor error messages, and
is annoying to maintain as the language gets more complex.

This first change is a mostly-faithful port of the yacc parser, though
it produces some different error messages and has some different opinions
about what source code positions it reports for different nodes in the
AST.

This is a hand-written recursive descent parser, so it uses native
recursion to process recursive expression structures. For the relatively-
simple expressions that are used with HIL this should not pose any
problem and it makes the parser control flow easier to follow.