Allow ast.Pos to have an optional filename
HIL expressions are often embedded in other files. A sufficiently-
sophisticated caller can pass ine the location where the HIL expression
starts relative to the broader file context, thus giving the user better
error diagnostics if a problem is encountered during either parsing or
evaluation.

This is entirely optional and will be ignored if empty.