Parse indexing [] as a postfix operator
Previously indexing was treated as a part of variable access, which meant
it couldn't be applied to non-variable expressions such as the result of
a function.

Now indexing is treated as a postfix operator in its own right, meaning
that it can follow any expression term.

This change is backward-compatible because a postfix operator following
a variable access produces an AST identical to the previous, and other
uses of brackets were previously a syntax error.

This doesn't change the type checking or evaluation behavior, so this change
alone doesn't actually make indexing of function return values work, it just
changes it from being a parse error to being a type checker error. The
short-term value here is to enable a more helpful error message for this
scenario, rather than the raw syntax error we had previously. A later
improvement to the type system could then make this work for real.