More user-friendly error for indexing non-variables
Due to limitations of HIL's type system we are currently only able to do
static type checking when indexing is applied directly to variables. We
can't, for example, type check an expression like split(" ", foo)[0]
because we only know that split returns a list, and not what it's a
list *of*.

We will hopefully support this later, but for now we can have an error
message that explicitly describes the problem.