LiteralNode factory function
Sometimes a caller wants to return a "synthetic" literal based on some
Go value it is holding, such as a default placeholder value that can
optionally be overridden by a HIL expression.

This new function helps callers do this robustly, selecting the
appropriate value for the Typex attribute based on the dynamic type of
the provided value.

For now this supports only primitive types. Later it could be interesting
to support recursive construction of lists and maps from Go slices and
maps, but for now we keep this relatively simple.