Type checking and evaluation of comparison operators
Equality and inequality apply to all of the primitive types, while the
<, >, <= and >= operators apply only to the numeric types.

This also implicitly implements the logical NOT operator (!) because the
parser represents it as an equality comparison with false.

Just as with the other arithmetic operations, these are transformed into
calls to builtin functions, with one function per type that supports
comparison operations.