parser+ast: new conditional expression node and associated parsing
This mimics the (.. ? .. : ..) from C, which has since been mimicked in
several other languages and is thus likely to be familiar to many users.

It is a ternary operator, but its parsing -- like C, again -- treats it
as a funny sort of binary operator where the operand is the middle
expression and the wrapping punctuation.

A new AST node type was introduced, rather than continuing to stretch the
meaning of the Arithmetic node type, because the type checking rules
and evaluation steps will be quite different than the arithmetic
operators, and the "type transparency" of the two output expressions will
make it it inconvenient to implement by mapping to a built-in function as
we do for the arithmetic operations.