Model container types in the type system
Previously Type was just an enumeration of simple types, but that meant
that there was no way for the type system to represent the element type
of a list or a map value.

Here we change Type to instead be an interface, and then provide separate
concrete types for primitives, lists, and maps. This initial patch
doesn't actually change any behavior to *use* the list and map element
types, but rather just creates a place for the element types to live and
then generally ignores it.

This intentionally preserves the convenient ability to compare types by
the standard equality operator == and treat the primitive types as
opaque values accessed via globals. Thus code dealing with the primitive
types is generally unchanged by this, but this is a breaking change for
code that deals with list and map types, and it is the updates for this
that make up the bulk of this change.

The most interesting part of this change is in ast/type.go where the new
type system is defined as a selection of Go types.