Type check indexing without assuming VariableAccess
Since the old type system could not represent the element type of a list
or map, the type check for the index operator [] previously required that
its operand be a variable access so that the variable's value could be
used to infer the element type.

We are now able to represent element types directly in the type system,
so we can eliminate this constraint and determine the element type based
only on the type information, without needing to inspect runtime data.