Parse indexing as a distinct postfix operator
Previously we treated indexing as an extension of variable access syntax
for parsing purposes. Now we treat it as a separate operator that can
appear after any term that returns a list or map of something.

This also allows us to include a test for the index type checking that
was generalized in the previous commit, since the type checker tests
depend on the parser.