Evaluate lists and maps that do not come from variables
Previously, due to historical constraints with the type system, we allowed
indexing only of variables directly. The type system and parser are now
generalized to support arbitrary indexing, so this completes the picture
making indexing on non-variable expressions return the expected value.

In particular, this allows indexing arrays of arrays and other such
nested data structures, and allows indexing into the results of functions
that return lists and maps.