Allow functions to operate generically on lists and maps
Now that lists and map types are parameterized by their element types we
expect that all collection values will carry a concrete element type.
However, some functions operate generically on collections of any type,
such as Terraform's "concat" function which should be able to accept
lists of any type (as long as they all match) and return a list of that
same type.

To support this we allow functions to optionally define their own type
checking logic, in addition to the simple type checking built in to HIL.
The use of TypeAny as a parameter type wildcard is extended to allow
functions to accept TypeList{TypeAny} and TypeMap{TypeAny} to express
generic list and map support, and then the function declaration can use
the new ReturnTypeFunc field to explain to HIL how the given argument
types map to a particular return type.

Returning to the "concat" example, this function would be defined as
being Variadic with a type of TypeList{TypeAny}, and then its
ReturnTypeFunc would check that the ElementType of all of the given lists
is identical and indicate that its return type is that ElementType.