Use correct element types when converting Go values to HIL variables
The functions in convert.go allow conversion between Go values and HIL
variable instances and vice-versa. This is used by calling applications
such as Terraform to integrate with other code that is not aware of HIL.

Now that our collection types have specified element types, we must
populate these correctly when converting from Go values. This problem
is simplified because these conversion functions use only the collection
types and the string type, and thus we can assume that we'll always end
up eventually making a string variable.