Give function implementations access to return type
Now that we support computed return types for functions, it's convenient
to retain the computed return type and pass it in to the function's
implementation so that it doesn't need to repeat the logic to determine
the result type when working with collections generically.

To support this we add a new CallbackTyped attribute to Function which
can be used instead of Callback when knowing the return type is desired.
A new AST node CallTyped is used to augment a Call node with its
computed result type from the type check phase, which can then be used
in the eval phase.