eval: check for unknowns in output before we begin concatenating
When the type checker finds unknowns it will generally skip any AST
rewriting it would normally do to insert type conversions, so if the
output result contains at least one unknown then it's possible that our
results would not all be of type string as we would usually expect.

To guard against this, we check first if there are any unknowns and just
return early, before we even start building our result buffer.